Create README.md
This was an exercise I went through this course by Andrew Ba Tran https://learn.r-journalism.com/en/introduction/ . He works at the Washington Post and the Knight School of Journalism had sponsored the course. The course is fully open to public now and it was actually a fun course to go through. In the last section (7) he goes through how to put your work on github and this is what I followed!